+++ pkg/vm/engine/tae/db/gc/checkpoint.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package gc2
+package gc
 
 import (
 	"context"
